chore(Dockerfile): update base image to golang:1.20.6-bullseye and debian:11.7-slim
The base image for the Dockerfile has been updated to
golang:1.20.6-bullseye and debian:11.7-slim to ensure compatibility and
use the latest versions of the dependencies. The Dockerfile now includes
instructions to build a statically linked binary by setting
CGO_ENABLED=0, GOOS=linux, and GOARCH=amd64 environment variables. The
upx package is installed to compress the binary. This improves the
portability and efficiency of the resulting Docker image.