ci(ci.yml): update Docker workflow to use docker-publish from main branch and specify image and Dockerfile
The GitHub Actions CI configuration now uses the `docker-publish.yaml`
workflow from the `main` branch instead of `docker-build-push.yaml` from
the `develop` branch. Additionally, specifying the
`image` and `Dockerfile` explicitly in the workflow allows for better
control and clarity over the Docker build process, ensuring that the
correct image and Dockerfile are used during the build.